refactor(administration): fold the transform-script wrapper chain into one file
`transform-script.ts` was a three-line barrel re-exporting from
`transform-script-implementation.ts`, which in turn only called
`build-composition-api-script.ts` to repackage three fields. Nothing else
imported the two middle files, so the whole chain collapses into
`transform-script.ts` holding the real implementation.
